change hiv_healthseekingbehaviour_cap and tb_healthseekingbehaviour_cap to type=INT otherwise condition doesn't work correctly

+++ src/tlo/methods/hiv.py
@@ -384,7 +384,7 @@ class Hiv(Module):
             "probability of death if aids and tb, person on treatment for tb",
         ),
         "hiv_healthseekingbehaviour_cap": Parameter(
-            Types.REAL,
+            Types.INT,
             "number of repeat visits assumed for healthcare services",
         ),
     }
